+++ lib/hand_landmarker_bindings.dart
@@ -133,7 +133,7 @@ class MyHandLandmarker extends jni$_.JObject {
 
   static final _id_detectFromYuv = _class.instanceMethodId(
     r'detectFromYuv',
-    r'(Ljava/nio/ByteBuffer;Ljava/nio/ByteBuffer;Ljava/nio/ByteBuffer;IIIIIIII)Ljava/lang/String;',
+    r'(Ljava/nio/ByteBuffer;Ljava/nio/ByteBuffer;Ljava/nio/ByteBuffer;IIIIII)Ljava/lang/String;',
   );
 
   static final _detectFromYuv = jni$_.ProtectedJniExtensions.lookup<
@@ -146,8 +146,6 @@ class MyHandLandmarker extends jni$_.JObject {
                         jni$_.Pointer<jni$_.Void>,
                         jni$_.Pointer<jni$_.Void>,
                         jni$_.Pointer<jni$_.Void>,
-                        jni$_.Int32,
-                        jni$_.Int32,
                         jni$_.Int32,
                         jni$_.Int32,
                         jni$_.Int32,
@@ -167,11 +165,9 @@ class MyHandLandmarker extends jni$_.JObject {
               int,
               int,
               int,
-              int,
-              int,
               int)>();
 
-  /// from: `public fun detectFromYuv(yBuffer: java.nio.ByteBuffer, uBuffer: java.nio.ByteBuffer, vBuffer: java.nio.ByteBuffer, width: kotlin.Int, height: kotlin.Int, yRowStride: kotlin.Int, uvRowStride: kotlin.Int, uvPixelStride: kotlin.Int, rotation: kotlin.Int, conversionMode: kotlin.Int, jpegQuality: kotlin.Int): kotlin.String`
+  /// from: `public fun detectFromYuv(yBuffer: java.nio.ByteBuffer, uBuffer: java.nio.ByteBuffer, vBuffer: java.nio.ByteBuffer, width: kotlin.Int, height: kotlin.Int, yRowStride: kotlin.Int, uvRowStride: kotlin.Int, uvPixelStride: kotlin.Int, rotation: kotlin.Int): kotlin.String`
   /// The returned object must be released after use, by calling the [release] method.
   jni$_.JString detectFromYuv(
     jni$_.JByteBuffer byteBuffer,
@@ -183,8 +179,6 @@ class MyHandLandmarker extends jni$_.JObject {
     int i3,
     int i4,
     int i5,
-    int i6,
-    int i7,
   ) {
     final _$byteBuffer = byteBuffer.reference;
     final _$byteBuffer1 = byteBuffer1.reference;
@@ -200,9 +194,7 @@ class MyHandLandmarker extends jni$_.JObject {
             i2,
             i3,
             i4,
-            i5,
-            i6,
-            i7)
+            i5)
         .object<jni$_.JString>(const jni$_.$JString$Type$());
   }
 }